+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## 3.4.0
 - Update queries for gamefic-3.7.0
-- Clips
+- Bound itemize methods for look actions
 
 ## 3.3.0 - October 5, 2024
 - Include Pronoun and Give in Standard